requirements.txt
aiofiles==24.1.0
aiolimiter==1.1.0
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.0
anytree==2.12.1
asttokens==2.4.1
attrs==24.2.0
autograd==1.7.0
azure-common==1.1.28
azure-core==1.31.0
azure-identity==1.18.0
azure-search-documents==11.5.1
azure-storage-blob==12.23.1
beartype==0.18.5
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
datashaper==0.0.49
decorator==5.1.1
deprecation==2.1.0
devtools==0.12.2
diskcache==5.6.3
distro==1.9.0
environs==11.0.0
executing==2.1.0
fonttools==4.54.1
future==1.0.0
gensim==4.3.3
gitdb==4.0.11
GitPython==3.1.43
graphrag==0.3.6
graspologic==3.4.1
graspologic-native==1.2.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
hyppo==0.4.0
idna==3.10
isodate==0.6.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
json_repair==0.28.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
lancedb==0.13.0
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
mdurl==0.1.2
msal==1.31.0
msal-extensions==1.2.0
narwhals==1.9.1
networkx==3.3
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
openai==1.51.0
overrides==7.7.0
packaging==24.1
pandas==2.2.3
patsy==0.5.6
pillow==10.4.0
portalocker==2.10.1
POT==0.9.4
protobuf==5.28.2
py==1.11.0
pyaml-env==1.2.1
pyarrow==15.0.2
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.9.0
pylance==0.17.0
pynndescent==0.5.13
pyparsing==3.1.4
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
retry==0.9.2
rich==13.9.2
rpds-py==0.20.0
scikit-learn==1.5.2
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
smart-open==7.0.5
smmap==5.0.1
sniffio==1.3.1
statsmodels==0.14.4
streamlit==1.39.0
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.7.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
umap-learn==0.5.6
urllib3==2.2.3
wrapt==1.16.0
